refactor: refactor imports to common/domain #64
