+++ backend/src/main/java/com/example/backend/dashboard/repository/CaseRepository.java
@@ -1,7 +1,7 @@
 package com.example.backend.dashboard.repository;
 
-import com.example.backend.dashboard.domain.CaseEntity;
-import com.example.backend.dashboard.domain.CaseEntity.CaseState;
+import com.example.backend.common.domain.CaseEntity;
+import com.example.backend.common.domain.CaseEntity.CaseState;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
