+++ backend/src/main/java/com/example/backend/user/dto/UserResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class UserResponseDto {
-    private int officeId;
+    private Integer officeId;
     private String name;
     private String profileUrl;
 }
