+++ backend/src/main/java/com/example/backend/user/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.example.backend.user.repository;
 
-import com.example.backend.user.domain.UserEntity;
+import com.example.backend.common.domain.PoliceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-    Optional<UserEntity> findByUserIdAndPassword(String userId, String password);
+public interface UserRepository extends JpaRepository<PoliceEntity, Integer> {
+    Optional<PoliceEntity> findByUserIdAndPassword(String userId, String password);
 }
