+++ backend/src/main/java/com/example/backend/user/service/UserService.java
@@ -1,8 +1,8 @@
 package com.example.backend.user.service;
 
+import com.example.backend.common.domain.PoliceEntity;
 import com.example.backend.user.dto.UserRequestDto;
 import com.example.backend.user.dto.UserResponseDto;
-import com.example.backend.user.domain.UserEntity;
 import com.example.backend.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,10 +24,10 @@ public class UserService {
     @Transactional(readOnly = true)  // 이 메서드는 읽기 전용으로 실행됨 (성능 최적화)
     public UserResponseDto authenticate(UserRequestDto requestDto) {
         // requestDto에서 받은 userId와 password를 기반으로 사용자 정보를 조회
-        Optional<UserEntity> user = userRepository.findByUserIdAndPassword(requestDto.getUserId(), requestDto.getPassword());
+        Optional<PoliceEntity> user = userRepository.findByUserIdAndPassword(requestDto.getUserId(), requestDto.getPassword());
 
         // 조회한 결과가 있으면 UserResponseDto 객체로 변환하여 반환하고, 없으면 예외 던짐
-        return user.map(value -> new UserResponseDto(value.getOfficeId(), value.getName(), value.getProfileUrl()))  // 사용자 이름을 포함한 응답 생성
+        return user.map(value -> new UserResponseDto(value.getOffice().getId(), value.getName(), value.getProfileUrl()))  // 사용자 이름을 포함한 응답 생성
                 .orElseThrow(() -> new IllegalArgumentException("Invalid credentials"));  // 사용자 정보가 없으면 예외 처리
     }
